refactor(transforms): extract shared prop-summarization toolkit (DIR-F1)
The `SummarizedProp` type, `getSummarizedProps`, `renderAsStatic`,
`getSummarizedPropValue`, `removePropsByName`, and `isExactKey` were
duplicated byte-identical between `input-text-area-transform.ts` and
`select-transform.ts`. Audit DIR-F1 flagged the ~85-LOC duplication as
a silent-drift hazard — particularly on the prop-name matching
contract (`isExactKey`), where each copy's docblock had drifted in
wording even though the body matched.

Lift the toolkit to `transforms/_shared-props.ts`. The underscore
prefix mirrors the existing convention for transform-internal modules
(no package barrel re-exports it). Both consumers import the symbols
they need; the local definitions are deleted.

Net: −129 LOC across the two transform files. No behaviour change —
the shared module is byte-for-byte the same shape both copies
implemented; the per-transform test suites stay green unchanged.
